Removed Center from all agents.

+++ Tellma.Database.Application/dal/Functions/dal.fn_Account_Center_Currency_Agent_Resource_NotedDate__Balance.sql
@@ -6,7 +6,7 @@
 	@ResourceId	INT,
 	@NotedDate	DATE
 )
-RETURNS INT
+RETURNS DECIMAL (19, 4)
 AS
 BEGIN
 	RETURN (
